+++ src/php/Prose/FromProcessesTable.php
@@ -87,6 +87,6 @@ class FromProcessesTable extends Prose
      */
     public function getDetailsForPid($pid)
     {
-        return fromRuntimeTable($this->tableName)->getDetails($pid);
+        return fromRuntimeTable($this->tableName)->getItem($pid);
     }
 }